+++ data_nature_importer/__manifest__.py
@@ -4,8 +4,8 @@
     "author": "BAKEUP",
     "website": "https://www.bakeup.org",
     "category": "product",
-    "version": "16.0.1.0.2",
-    "depends": ["product", "queue_job"],
+    "version": "16.0.1.2.0",
+    "depends": ["product", "queue_job", "website_sale"],
     "data": [
         "security/ir.model.access.csv",
         "views/product_template_views.xml",